# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI智能助手系统",
    "summary": "AI智能助手系统API",
    "description": "\n## AI智能助手系统后端API\n\n基于FastAPI和LangChain 1.0框架构建的企业级AI智能助手系统，提供以下核心功能：\n\n### 核心功能\n\n* **智能对话** - 支持多轮对话、流式响应、对话历史管理\n* **RAG知识库** - 文档上传、向量化、智能检索问答\n* **Agent智能代理** - 工具调用、任务执行、推理过程可视化\n* **用户管理** - 注册登录、JWT认证、配额管理\n* **系统监控** - 使用统计、健康检查、Prometheus指标\n\n### 技术栈\n\n* **Web框架**: FastAPI 0.104+\n* **AI框架**: LangChain 1.0\n* **数据库**: MySQL 8.0, Redis 7.0, Chroma\n* **LLM**: 通义千问 (Tongyi Qianwen)\n\n### 认证方式\n\n大部分API端点需要JWT认证。在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 速率限制\n\n* 登录接口: 5次/分钟\n* 普通API: 100次/分钟\n* LLM调用: 20次/分钟\n\n### 错误响应格式\n\n所有错误响应遵循统一格式：\n```json\n{\n    \"error_code\": \"1001\",\n    \"message\": \"错误描述\",\n    \"request_id\": \"uuid\"\n}\n```\n    ",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "聊天"
        ],
        "summary": "发送消息（非流式）",
        "description": "发送消息到对话，返回AI回复。",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": [
          "聊天"
        ],
        "summary": "发送消息（流式）",
        "description": "发送消息到对话，使用SSE协议流式返回AI回复。",
        "operationId": "stream_chat_api_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatConfig": {
        "properties": {
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "description": "温度参数，控制输出随机性",
            "default": 0.7
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 4000.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "description": "最大输出token数",
            "default": 2000
          },
          "mode": {
            "$ref": "#/components/schemas/ChatModeEnum",
            "description": "对话模式",
            "default": "normal"
          }
        },
        "type": "object",
        "title": "ChatConfig",
        "description": "对话配置模型\n\n用于配置对话的模型参数。"
      },
      "ChatModeEnum": {
        "type": "string",
        "enum": [
          "normal",
          "professional",
          "creative"
        ],
        "title": "ChatModeEnum",
        "description": "对话模式枚举"
      },
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id",
            "description": "对话ID",
            "examples": [
              1,
              42
            ]
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Content",
            "description": "消息内容",
            "examples": [
              "你好，请介绍一下Python",
              "什么是机器学习？"
            ]
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "对话配置"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "content"
        ],
        "title": "ChatRequest",
        "description": "聊天请求模型\n\n用于发送消息到对话。",
        "examples": [
          {
            "config": {
              "max_tokens": 2000,
              "mode": "normal",
              "temperature": 0.7
            },
            "content": "你好，请介绍一下Python",
            "conversation_id": 1
          },
          {
            "config": {
              "max_tokens": 1500,
              "mode": "professional",
              "temperature": 0.3
            },
            "content": "帮我写一个快速排序算法",
            "conversation_id": 2
          }
        ]
      },
      "ChatResponse": {
        "properties": {
          "message_id": {
            "type": "integer",
            "title": "Message Id",
            "description": "AI回复消息ID"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "AI回复内容"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "description": "消耗的token数量",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "content"
        ],
        "title": "ChatResponse",
        "description": "聊天响应模型（非流式）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}